Add logger to SummaryManager; move ITreeEntry implementations to core-utils package
