+++ packages/runtime/container-runtime/src/index.ts
@@ -4,4 +4,3 @@
  */
 
 export * from "./containerRuntime";
-export * from "./utils";
